rnn/requirements.txt
matplotlib
seaborn
tqdm
pandas
numpy
tensorboardX
tensorflow
pytorch
